Metatask apply/remove leave the praxis service too
Second cut of #1391. `_check_metatask_eligibility`, `apply_metatask` and
`remove_metatask` move to `services/praxis_metatask.py`; `routers/praxes.py` is
their only consumer and repoints. No test imported them.

This module cannot be a leaf the way `praxis_out` is — both ops load and re-read
the aggregate through `services.praxis.get_praxis` — so the edge runs
`praxis_metatask` -> `praxis` and never back, and the router imports directly
rather than through a re-export that would invert it.

Deleted rather than moved: `services/praxis.py`'s own `ALBESCENT_FACTION_SLUG`
literal, a third copy of a constant `services/character.py` already owns and
`services/faction_service.py` already re-exports. The new module reads it from
`faction_service`, which it imports anyway for `faction_permits`.
