+++ backend/services/praxis.py
@@ -17,7 +17,6 @@ from sqlalchemy.orm import selectinload
 from game_config import CURRENT_ERA, EraConfig
 from models.character import Character
 from models.flag import Flag, FlagReason, stored_flag_reason
-from models.meta_task import PraxisMetaTask
 from models.praxis import (
     MediaItem,
     ModerationStatus,
@@ -43,7 +42,6 @@ from services.character_stats import (
 )
 from services.faction_service import faction_permits
 from services.media import process_and_save_media
-from services.meta_task import metatask_cap_for_level
 from services.era import (
     get_current_era_row,
     get_current_era_row_safe,
@@ -52,11 +50,6 @@ from services.era import (
 )
 from services.level_jump import available_level_reach, consume_level_jump
 from models.duel import Duel, DuelStatus
-
-
-# No EVERYMEN_FACTION_SLUG: the Double Dipper ability is a FactionConfig field
-# read via multi_membership_faction_slugs, not a slug to compare against (#1359).
-ALBESCENT_FACTION_SLUG = "albescent"
 
 
 # ---------------------------------------------------------------------------
@@ -1675,164 +1668,10 @@ async def moderate_praxis(
     return await get_praxis(praxis_id, session)
 
 
-# ---------------------------------------------------------------------------
-# Metatasks
-# ---------------------------------------------------------------------------
-
-
-def _check_metatask_eligibility(
-    character: Character,
-    task: Task,
-    character_level: int,
-    era: EraConfig,
-) -> Optional[str]:
-    """Return a 403 reason string if this character can't apply ``task``, else None."""
-    # Albescent bypasses the level gate (its charter); everyone else must meet
-    # metatask_apply_level. Metatasks are faction-open, so the seam
-    # `faction_permits` (ADR-0029, #171) currently permits every faction — the
-    # call is retained so a future faction rule is inherited here automatically.
-    if character.faction_slug == ALBESCENT_FACTION_SLUG:
-        return None
-    if character_level < era.metatask_apply_level:
-        return (
-            f"Must be level {era.metatask_apply_level} or above "
-            "to apply metatasks."
-        )
-    if not faction_permits(character, task, era):
-        return "This metatask belongs to a different faction."
-    return None
-
-
-async def apply_metatask(
-    praxis_id: int,
-    task_id: int,
-    character_id: int,
-    session: AsyncSession,
-    era: EraConfig = CURRENT_ERA,
-) -> Praxis:
-    """Attach a metatask (Task with task_type=metatask) to a praxis.
-
-    Gates (R.9):
-    - The task must be ``TaskType.metatask`` (else 400).
-    - The applying character must be a member of the praxis (else 403).
-    - The praxis must be ``in_progress`` (else 422).
-    - Level gate: at least ``era.metatask_apply_level`` (Albescent bypasses).
-      Metatasks are faction-open — any faction may apply any
-      faction's metatask.
-    - Quantity cap: at most ``metatask_cap_for_level(level, era)`` metatasks on
-      one praxis (else 422).
-    """
-    praxis = await get_praxis(praxis_id, session)
-    task = await session.get(Task, task_id)
-    if task is None:
-        raise HTTPException(status_code=404, detail="Task not found.")
-    if task.task_type != TaskType.metatask:
-        raise HTTPException(
-            status_code=400,
-            detail="Only tasks with task_type='metatask' can be applied as metatasks.",
-        )
-
-    member_ids = {member.character_id for member in praxis.members}
-    if character_id not in member_ids:
-        raise HTTPException(
-            status_code=403,
-            detail="Only members of this praxis can apply metatasks.",
-        )
-
-    if praxis.status != PraxisStatus.in_progress:
-        raise HTTPException(
-            status_code=422,
-            detail="Metatasks can only be applied to in-progress praxes.",
-        )
-
-    character = await session.get(Character, character_id)
-    if character is None:
-        raise HTTPException(status_code=404, detail="Character not found.")
-
-    era_row = await get_current_era_row(session)
-    stats = await get_or_create_stats(session, character_id, era_row.id)
-
-    eligibility_error = _check_metatask_eligibility(character, task, stats.level, era)
-    if eligibility_error is not None:
-        raise HTTPException(status_code=403, detail=eligibility_error)
-
-    # Reject duplicate links up front — a metatask can only be applied once
-    # to the same praxis.
-    existing = await session.execute(
-        select(PraxisMetaTask).where(
-            PraxisMetaTask.praxis_id == praxis_id,
-            PraxisMetaTask.task_id == task_id,
-        )
-    )
-    if existing.scalar_one_or_none() is not None:
-        raise HTTPException(
-            status_code=409,
-            detail="This metatask is already applied to the praxis.",
-        )
-
-    # Quantity cap: how many metatasks this praxis may hold rises with the
-    # applying character's level (metatasks_per_praxis_max_level).
-    cap = metatask_cap_for_level(stats.level, era)
-    current_count = await session.scalar(
-        select(func.count())
-        .select_from(PraxisMetaTask)
-        .where(PraxisMetaTask.praxis_id == praxis_id)
-    )
-    if current_count >= cap:
-        raise HTTPException(
-            status_code=422,
-            detail=(
-                f"This praxis already holds the maximum of {cap} "
-                f"metatask{'s' if cap != 1 else ''} at your level."
-            ),
-        )
-
-    session.add(PraxisMetaTask(praxis_id=praxis_id, task_id=task_id))
-    await session.flush()
-
-    await recalculate_members_stats(praxis, session, era)
-    return await get_praxis(praxis_id, session)
-
-
-async def remove_metatask(
-    praxis_id: int,
-    task_id: int,
-    character_id: int,
-    session: AsyncSession,
-    era: EraConfig = CURRENT_ERA,
-) -> Praxis:
-    """Remove a metatask from a praxis. Any praxis member can remove."""
-    praxis = await get_praxis(praxis_id, session)
-
-    member_ids = {member.character_id for member in praxis.members}
-    if character_id not in member_ids:
-        raise HTTPException(
-            status_code=403,
-            detail="Only members of this praxis can remove metatasks.",
-        )
-
-    result = await session.execute(
-        select(PraxisMetaTask).where(
-            PraxisMetaTask.praxis_id == praxis_id,
-            PraxisMetaTask.task_id == task_id,
-        )
-    )
-    link = result.scalar_one_or_none()
-    if link is None:
-        raise HTTPException(status_code=404, detail="Metatask is not applied to this praxis.")
-
-    await session.delete(link)
-    await session.flush()
-
-    await recalculate_members_stats(praxis, session, era)
-    return await get_praxis(praxis_id, session)
-
-
 __all__ = [
     "active_member_task_ids",
     "active_member_task_ids_subquery",
     "allowed_praxis_modes",
-    "apply_metatask",
     "cancel_pending_publish_on_edit",
     "can_flag_praxis",
     "can_submit_praxis_for_task",
@@ -1856,7 +1695,6 @@ __all__ = [
     "VotedFilter",
     "moderate_praxis",
     "multi_membership_faction_slugs",
-    "remove_metatask",
     "respond_to_invite",
     "SignupDenialReason",
     "SignupEligibility",
